refactor: reorganize authentication forms and enhance UI consistency
- Updated imports in authentication forms to use a new centralized structure, improving code organization.
- Replaced deprecated components with new form elements for better semantic structure and user experience.
- Introduced new AccountPrompt component for consistent messaging across SignIn, SignUp, ForgotPassword, and ResetPassword forms.
- Enhanced error handling and user feedback in forms, ensuring clarity and consistency in messaging.
- Added new forms for email verification, password reset, and forgot password functionalities, streamlining user interactions.
