+++ src/components/ui/index.tsx
@@ -1,10 +1,10 @@
+export * from './account-prompt';
 export * from './button';
 export * from './card';
 export * from './form';
 export * from './input';
 export * from './label';
 export * from './password-input';
-export * from './textarea';
 export * from './icon-badge';
 export * from './container';
 export * from './separator';
